Fix Avalara calculation on digital orders
